+++ internal/usergroup/add_unsupported.go
@@ -1,6 +1,6 @@
 //go:build !linux
 
-package idtools // import "github.com/docker/docker/pkg/idtools"
+package usergroup
 
 import "fmt"
