+++ internal/usergroup/utils_unix.go
@@ -1,6 +1,6 @@
 //go:build !windows
 
-package idtools // import "github.com/docker/docker/pkg/idtools"
+package usergroup
 
 import (
 	"fmt"
